requirements.txt
boto3==1.11.14
python-pptx==0.6.18
